Add Batched support in TriMesh
