+++ src/metrics/mesh.jl
@@ -2,19 +2,23 @@ export laplacian_loss, edge_loss
 
 function laplacian_loss(m::TriMesh)
     # TODO: There will be some changes when migrating to batched format
-    my_ignore() do
-        L = get_laplacian_sparse(m)
+    L = my_ignore() do
+         get_laplacian_packed(m)
     end
-    L = L * m.vertices
+    verts = get_verts_packed(m)
+    L = L * verts
     L = _norm(L; dims = 2)
     return sum(L)
 end
 
 function _edge_loss2(m::TriMesh, target_length::Number=0.0f0)
     #TODO: This is different approach to calculate edge_loss, remove one approach
-    p1 = m.vertices[m.faces[:,1],:]
-    p2 = m.vertices[m.faces[:,2],:]
-    p3 = m.vertices[m.faces[:,3],:]
+    verts = get_verts_packed(m)
+    faces = get_faces_packed(m)
+
+    p1 = verts[faces[:,1],:]
+    p2 = verts[faces[:,2],:]
+    p3 = verts[faces[:,3],:]
 
     e1 = p2-p1
     e2 = p3-p2
@@ -31,9 +35,10 @@ end
 
 function edge_loss(m::TriMesh, target_length::Number=0.0f0)
     #TODO: will change changing to batched format
-    edges = Zygote.nograd(get_edges(m))
-    v1 = m.vertices[edges[:,1],:]
-    v2 = m.vertices[edges[:,2],:]
+    verts = get_verts_packed(m)
+    edges = Zygote.nograd(get_edges_packed(m))
+    v1 = verts[edges[:,1],:]
+    v2 = verts[edges[:,2],:]
     el =  (_norm(v1-v2; dims=2) .- Float32(target_length)) .^ 2
     loss = mean(el)
     return loss
