+++ src/transforms/mesh_func.jl
@@ -7,9 +7,9 @@ import Distributions
 
 Uniformly samples `num_samples` points from the surface of TriMesh `m`.
 
-`returns_normals` is optional keyword argument, to make returns normals
+`returns_normals` is optional keyword argument, to returns normals
 from respective faces of samples. `eps` is optional keyword argument for
-a small number to prevent division by zero for small surface areas. 
+a small number to prevent division by zero for small surface areas.
 
 ### Examples:
 julia> m = load_trimesh("teapot.obj")
@@ -18,31 +18,55 @@ julia> points, normals = sample_points(m, 5000; returns_normals=true)
 
 """
 function sample_points(
-    m::TriMesh,
+    m::TriMesh{T,R},
     num_samples::Int = 5000;
     returns_normals::Bool = false,
     eps::Number = EPS,
-)
-    face_areas, face_normals = compute_face_areas(m; compute_normals = returns_normals)
-    face_areas_prob = Float64.(face_areas) ./ sum(Float64.(face_areas))
+)   where {T,R}
+    verts_padded = get_verts_padded(m)
+    faces_padded = get_faces_padded(m)
+    faces_areas_padded, faces_normals_padded = compute_faces_areas_padded(m; compute_normals = returns_normals)
+    
+        #(Fi,B)
     #TODO: condition for probvec fails in Float32
+    faces_areas_padded = Float64.(faces_areas_padded)
+    faces_areas_prob = faces_areas_padded ./ max.(sum(faces_areas_padded; dims=2), eps)
+    samples = Zygote.Buffer(T, num_samples, 3, m.N)
+    if returns_normals
+        samples_normals = Zygote.Buffer(T, num_samples, 3, m.N)
+    end
 
-    # face_areas_prob = face_areas ./ sum(face_areas)
-    # face_areas_prob = face_areas ./ max(sum(face_areas), eps)
-    dist = Distributions.Categorical(face_areas_prob)
-    sample_faces_idx = my_rand(dist, num_samples)
-    sample_faces_idx = Zygote.nograd(sample_faces_idx)
-    sample_faces = m.faces[sample_faces_idx, :]
-    v1 = m.vertices[sample_faces[:, 1], :]
-    v2 = m.vertices[sample_faces[:, 2], :]
-    v3 = m.vertices[sample_faces[:, 3], :]
-    (w1, w2, w3) = _rand_barycentric_coords(num_samples)
-    samples = (w1 .* v1) + (w2 .* v2) + (w3 .* v3)
+    for (i,_len) in enumerate(m._faces_len)
+        
+        dist = Distributions.Categorical(faces_areas_prob[1:_len, i])
+        sample_faces_idx = my_rand(dist, num_samples)
+        sample_faces_idx = Zygote.nograd(sample_faces_idx)
+        sample_faces = faces_padded[sample_faces_idx, :, i]
+        samples[:,:,i] = _sample_points(verts_padded[1:m._verts_len[i], 3, i], sample_faces, num_samples)
+        
+        if returns_normals
+            samples_normals[:,:,i] = faces_normals_padded[sample_faces_idx, :, i]
+        end
+    end
 
     if returns_normals
-        samples_normals = face_normals[sample_faces_idx, :]
-        return (samples, samples_normals)
+        return copy(samples), copy(samples_normals)
     end
+
+    return copy(samples)
+end
+
+function _sample_points(
+    verts::AbstractArray{T,2},
+    sample_faces::AbstractArray{R,2},
+    num_samples::Int
+)   where {T,R}
+
+    v1 = verts[sample_faces[:, 1], :]
+    v2 = verts[sample_faces[:, 2], :]
+    v3 = verts[sample_faces[:, 3], :]
+    (w1, w2, w3) = _rand_barycentric_coords(num_samples)
+    samples = (w1 .* v1) + (w2 .* v2) + (w3 .* v3)
     return samples
 end
 
@@ -58,8 +82,8 @@ end
 """
     normalize!(m::TriMesh)
 
-Normalize the TriMesh `m` with mean centered at origin and unit standard deviation
-and overwrite the m with normalized TriMesh.
+Normalize each mesh in TriMesh `m` with mean centered at origin and unit standard deviation
+and overwrite the `m` with normalized TriMesh.
 
 See also: [`normalize`](@ref)
 
@@ -71,15 +95,17 @@ julia> normalize!(m)
 ```
 """
 function normalize!(m::TriMesh)
-    centroid = mean(m.vertices; dims = 1)
-    m.vertices = (m.vertices .- centroid) ./ (std(m.vertices, mean = centroid, dims = 1) .+ EPS)
+    verts_padded = get_verts_padded(m)
+    centroid = mean(verts_padded; dims = 1)
+    verts_padded = (verts_padded .- centroid) ./ (std(verts_padded, mean = centroid, dims = 1) .+ EPS)
+    m._verts_padded = _padded_to_packed(verts_padded, m._verts_len)
     return m
 end
 
 """
     normalize(m::TriMesh)
 
-Normalize the TriMesh `m` with mean centered at origin and unit standard deviation
+Normalize each mesh in TriMesh `m` with mean centered at origin and unit standard deviation
 
 See also: [`normalize!`](@ref)
 
@@ -103,7 +129,7 @@ end
 Scale the TriMesh `m` by scaling factor `factor`
 and overwrite `m` with scaled TriMesh. If `factor`
 is array of size `(3, )`, then TriMesh will be scale
-by scaling factor of respective dimension.  
+by scaling factor of respective dimension.
 
 Scaling factor `factor` (each element in case of array)
 should be strictly greater than `0.0`.
@@ -119,14 +145,17 @@ julia> scale!(m, [1.0, 1.0, 1.0])
 """
 function scale!(m::TriMesh, factor::Float32)
     (factor > 0.0) || error("factor must be greater than 0.0")
-    lmul!(factor, m.vertices)
+    verts_packed = get_verts_packed(m)
+    lmul!(factor, verts_packed)
+    m._verts_packed = verts_packed
     return m
 end
 
 function scale!(m::TriMesh, factor::AbstractArray{Float32})
-    (size(factor) == (3,)) || error("factor must be (3, ), but instead got $(size(factor)) array")  
+    (size(factor) == (3,)) || error("factor must be (3, ), but instead got $(size(factor)) array")
     (factor .> 0.0) || error("factor must be greater than 0.0")
-    m.vertices = m.vertices .* reshape(factor, 1, 3)
+    verts_packed = get_verts_packed(m) .* reshape(factor, 1, 3)
+    m._verts_packed = verts_packed
     return m
 end
 
@@ -136,9 +165,9 @@ scale!(m::TriMesh, factor) = scale!(m, Float32.(factor))
     scale(m::TriMesh, factor::Number)
     scale(m::TriMesh, factor::AbstractArray{<:Number})
 
-Scale the TriMesh `m` by scaling factor `factor`. 
-If `factor` is array of size `(3, )`, then TriMesh 
-will be scaleby scaling factor of respective dimension. 
+Scale the TriMesh `m` by scaling factor `factor`.
+If `factor` is array of size `(3, )`, then TriMesh
+will be scaleby scaling factor of respective dimension.
 
 Scaling factor `factor` (each element in case of array)
 should be strictly greater than `0.0`.
@@ -178,7 +207,8 @@ julia> rotate!(m, rotmat)
 """
 function rotate!(m::TriMesh, rotmat::AbstractArray{Float32,2})
     size(rotmat) == (3, 3) || error("rotmat must be (3, 3) array, but instead got $(size(rotmat)) array")
-    m.vertices = m.vertices * rotmat
+    verts_packed = get_verts_packed(m) * rotmat
+    m._verts_packed = verts_packed
     return m
 end
 
@@ -211,8 +241,8 @@ end
     realign!(src::TriMesh, tgt::TriMesh)
     realign!(src::TriMesh, tgt_min::AbstractArray{<:Number,2}, tgt_max::AbstractArray{<:Number,2})
 
-Re-Align the TriMesh `src` with the axis aligned bounding box of TriMesh `tgt`
-and overwrite `src` with re-aligned TriMesh.
+Re-Align the TriMesh `src` with the axis aligned bounding box of first mesh in TriMesh `tgt`
+and overwrite `src` with re-aligned TriMesh. 
 
 See also: [`realign`](@ref)
 
@@ -224,9 +254,11 @@ julia> realign!(src, tgt)
 ```
 """
 function realign!(src::TriMesh, tgt_min::AbstractArray{Float32,2}, tgt_max::AbstractArray{Float32,2})
-    src_min = reshape(minimum(src.vertices, dims = 1), (1, :))
-    src_max = reshape(maximum(src.vertices, dims = 1), (1, :))
-    src.vertices = ((src.vertices .- src_min) ./ (src_max - src_min .+ EPS)) .* (tgt_max - tgt_min) .+ tgt_min
+    verts_padded = get_verts_padded(src)
+    src_min = reshape(minimum(verts_padded, dims = 1), (1, 3, :))
+    src_max = reshape(maximum(verts_padded, dims = 1), (1, 3, :))
+    verts_padded = ((verts_padded .- src_min) ./ (src_max - src_min .+ EPS)) .* (tgt_max - tgt_min) .+ tgt_min
+    src._verts_packed = _padded_to_packed(verts_padded, src._verts_len)
     return src
 end
 
@@ -234,8 +266,9 @@ realign!(src::TriMesh, tgt_min::AbstractArray{<:Number,2}, tgt_max::AbstractArra
     realign!(src, Float32.(tgt_min), Float32.(tgt_max))
 
 function realign!(src::TriMesh, tgt::TriMesh)
-    tgt_min = reshape(minimum(tgt.vertices, dims = 1), (1, :))
-    tgt_max = reshape(maximum(tgt.vertices, dims = 1), (1, :))
+    verts = get_verts_list(tgt)[1]
+    tgt_min = reshape(minimum(verts, dims = 1), (1, :))
+    tgt_max = reshape(maximum(verts, dims = 1), (1, :))
     realign!(src, tgt_min, tgt_max)
     return src
 end
@@ -244,7 +277,7 @@ end
     realign(src::TriMesh, tgt::TriMesh)
     realign(src::TriMesh, tgt_min::AbstractArray{<:Number,2}, tgt_max::AbstractArray{<:Number,2})
 
-Re-Align the TriMesh `src` with the axis aligned bounding box of TriMesh `tgt`.
+Re-Align the TriMesh `src` with the axis aligned bounding box of first mesh in TriMesh `tgt`.
 
 See also: [`realign`](@ref)
 
@@ -274,7 +307,7 @@ end
 Translate the TriMesh `m` by translating vector `vector`
 and overwrite `m` with translated TriMesh. If `vector`
 is a number, then TriMesh will be translated by same number
-in all dimension. 
+in all dimension.
 
 See also: [`translate`](@ref)
 
@@ -285,14 +318,13 @@ julia> translate!(m, 0.0)
 julia> translate!(m, [0.0, 0.0, 0.0])
 
 """
-function translate!(m::TriMesh, vector::Float32)
-    m.vertices = m.vertices .+ vector
-    return m
-end
-
+translate!(m::TriMesh, vector::Float32) = translate(m, fill(vector, (3,)))
+ 
 function translate!(m::TriMesh, vector::AbstractArray{Float32})
-    (size(vector) == (3,)) || error("vector must be (3, ), but instead got $(size(vector)) array")  
-    m.vertices = m.vertices .+ reshape(vector, 1, 3)
+    (size(vector) == (3,)) || error("vector must be (3, ), but instead got $(size(vector)) array")
+    verts_packed = get_verts_packed(m)
+    verts_packed = verts_packed .+ reshape(vector, 1, 3)
+    m._verts_packed = verts_packed
     return m
 end
 
@@ -302,9 +334,9 @@ translate!(m::TriMesh, vector) = translate!(m, Float32.(vector))
     translate(m::TriMesh, vector::Number)
     translate(m::TriMesh, vector::AbstractArray{<:Number})
 
-Translate the TriMesh `m` by translating vector `vector`. 
-If `vector` is a number, then TriMesh will be translated 
-by same number in all dimension. 
+Translate the TriMesh `m` by translating vector `vector`.
+If `vector` is a number, then TriMesh will be translated
+by same number in all dimension.
 
 See also: [`translate!`](@ref)
 
@@ -323,7 +355,7 @@ end
 """
     offset!(m::TriMesh, offset_verts::AbstractArray{<:Number,2})
 
-Add offset to the vertices of the TriMesh `m` by offset vertices `offset_verts`
+Add offset to the vertices of the TriMesh `m` by offset vertices `offset_verts_packed`
 and overwrite `m` with updated TriMesh.
 
 See also: [`offset`](@ref)
@@ -331,35 +363,37 @@ See also: [`offset`](@ref)
 ### Examples:
 ```julia
 julia> m = load_trimesh("teapot.obj")
-julia> offset_verts = ones(m.vertices)
+julia> offset_verts = ones(get_verts_packed(m))
 julia> offset!(m, offset_verts)
 
 """
 
-function offset!(m::TriMesh, offset_verts::AbstractArray{Float32, 2})
-    (size(offset_verts) == size(m.vertices)) || error("mesh and offset_verts size mismatch")
-    m.vertices = m.vertices + offset_verts
+function offset!(m::TriMesh, offset_verts_packed::AbstractArray{Float32, 2})
+    verts_packed = get_verts_packed(m)
+    (size(offset_verts_packed) == size(verts_packed)) || error("mesh and offset_verts size mismatch")
+    verts_packed += offset_verts_packed
+    m._verts_packed = verts_packed
     return m
 end
 
-offset!(m::TriMesh, offset_verts::AbstractArray{<:Number, 2}) = 
-    offset!(m, Float32.(offset_verts))
-    
-"""
-    offset(m::TriMesh, offset_verts::AbstractArray{<:Number,2})
+offset!(m::TriMesh, offset_verts_packed::AbstractArray{<:Number, 2}) =
+    offset!(m, Float32.(offset_verts_packed))
 
-Add offset to the vertices of the TriMesh `m` by offset vertices `offset_verts`.
+"""
+    offset(m::TriMesh, offset_verts_packed::AbstractArray{<:Number,2})
+
+Add offset to the vertices of the TriMesh `m` by offset vertices `offset_verts_packed`.
 
 See also: [`offset!`](@ref)
 
 ### Examples:
 ```julia
 julia> m = load_trimesh("teapot.obj")
-julia> offset_verts = ones(m.vertices)
+julia> offset_verts = ones(get_verts_packed(m))
 julia> m = offset(m, offset_verts)
 
 """
-function offset(m::TriMesh, offset_verts::AbstractArray{<:Number, 2})
+function offset(m::TriMesh, offset_verts_packed::AbstractArray{<:Number, 2})
     m = deepcopy(m)
-    offset!(m, offset_verts)
+    offset!(m, offset_verts_packed)
 end
